+++ bee-transaction/src/atomic/payload/signed_transaction/unsigned_transaction.rs
@@ -14,7 +14,7 @@ pub use crate::atomic::payload::signed_transaction::output::Output;
 use crate::atomic::payload::Payload;
 
 pub struct UnsignedTransaction {
-    pub input_count: u16,
+    pub input_count: u8,
     pub inputs: Vec<Input>,
     pub output_count: u16,
     pub outputs: Vec<Output>,
